feat: upload function setup
upload implementation
